Fix : 타입 선언 위치 수정
