--- client/src/components/login/LoginType.d.ts
@@ -1,15 +1,0 @@
-export interface LoginTypes {
-   username: string;
-   password: string;
-}
-
-interface SocialNum {
-   RRNConfirm: string;
-}
-
-export interface FindPasswordType {
-   [key: string]: string;
-   email: string;
-   question: string;
-   answer: string;
-}
