+++ client/src/components/signup/SignupForm.tsx
@@ -2,7 +2,7 @@ import styled from "styled-components";
 import { useForm, SubmitHandler } from "react-hook-form";
 import { useState } from "react";
 import SignupInput from "./SignupInput";
-import { SignupTypes } from "./SignupTypes";
+import { SignupTypes } from "./contentsArray";
 import SignupQuestion from "./SignupQuestion";
 import logo from "../../assets/img/logo2.png";
 import { signupPost } from "../../api/axios";
