--- client/src/components/signup/SignupTypes.d.ts
@@ -1,9 +1,0 @@
-export interface SignupTypes {
-   nickname: string;
-   email: string;
-   password: string;
-   passwordConfirm: string;
-   RRN: string;
-   question: string;
-   answer: string;
-}
